Initial version for BT82X & LVGL9.2.2

+++ Demo_lvgl/lvgl/tests/src/test_cases/test_demo_stress.c
@@ -1,5 +1,6 @@
 #if LV_BUILD_TEST
 #include "../lvgl.h"
+#include "../../lvgl_private.h"
 #include "../demos/lv_demos.h"
 
 #include "unity/unity.h"
@@ -26,7 +27,7 @@ void test_demo_stress(void)
 #endif
     /* loop once to allow objects to be created */
     loop_through_stress_test();
-    uint32_t mem_before = lv_test_get_free_mem();
+    size_t mem_before = lv_test_get_free_mem();
     /* loop 5 more times */
     for(uint32_t i = 0; i < 5; i++) {
         loop_through_stress_test();
